Development

+++ srv/src/controllers/members.controller.ts
@@ -18,13 +18,12 @@ import {ValidationPipe} from "../validation/validation.pipe";
 import {ValidationSchema} from "../validation/validation.schema";
 import {Tenant} from "../tenants/tenant.entity";
 import {JwtAuthGuard} from "../auth/jwt-auth.guard";
-import {ScopeGuard} from "../scopes/scope.guard";
-import {Scopes} from "../scopes/scopes.decorator";
-import {ScopeEnum} from "../scopes/scope.enum";
 import {User} from "../users/user.entity";
 import {Scope} from "../scopes/scope.entity";
 import {SecurityService} from "../scopes/security.service";
 import {ScopeService} from "../scopes/scope.service";
+import {Action} from "../scopes/actions.enum";
+import {ForbiddenException} from "../exceptions/forbidden.exception";
 
 @Controller('api/tenant')
 @UseInterceptors(ClassSerializerInterceptor)
@@ -40,14 +39,13 @@ export class MemberController {
     }
 
     @Get('/:tenantId/members')
-    @UseGuards(JwtAuthGuard, ScopeGuard)
-    @Scopes(ScopeEnum.TENANT_ADMIN, ScopeEnum.TENANT_VIEWER)
+    @UseGuards(JwtAuthGuard)
     async getTenantMembers(
         @Request() request,
         @Param('tenantId') tenantId: string
     ): Promise<User[]> {
         let tenant = await this.tenantService.findById(tenantId);
-        await this.securityService.contextShouldBeTenantViewer(request, tenant.domain);
+        this.securityService.check(request, Action.Read, tenant);
         let members = await this.usersService.findByTenant(tenant);
         for (const member of members) {
             member.scopes = await this.scopeService.getMemberScopes(tenant, member);
@@ -56,23 +54,21 @@ export class MemberController {
     }
 
     @Post('/:tenantId/member/:email')
-    @UseGuards(JwtAuthGuard, ScopeGuard)
-    @Scopes(ScopeEnum.TENANT_ADMIN)
+    @UseGuards(JwtAuthGuard)
     async addMember(
         @Request() request,
         @Param('tenantId') tenantId: string,
         @Param('email') email: string
     ): Promise<Tenant> {
         const user = await this.usersService.findByEmail(email);
-        let tenant = await this.tenantService.findById(tenantId);
-        await this.securityService.currentUserShouldBeTenantAdmin(request, tenant.domain);
+        const tenant = await this.tenantService.findById(tenantId);
+        this.securityService.check(request, Action.Update, tenant);
         await this.tenantService.addMember(tenantId, user);
         return tenant;
     }
 
     @Delete('/:tenantId/member/:email')
-    @UseGuards(JwtAuthGuard, ScopeGuard)
-    @Scopes(ScopeEnum.TENANT_ADMIN)
+    @UseGuards(JwtAuthGuard)
     async removeMember(
         @Request() request,
         @Param('tenantId') tenantId: string,
@@ -80,13 +76,16 @@ export class MemberController {
     ): Promise<Tenant> {
         const user = await this.usersService.findByEmail(email);
         let tenant = await this.tenantService.findById(tenantId);
-        await this.securityService.currentUserShouldBeTenantAdmin(request, tenant.domain)
+        this.securityService.check(request, Action.Update, tenant);
+        let securityContext = this.securityService.getUserSecurityContext(request);
+        if (securityContext.email === email) {
+            throw new ForbiddenException("cannot remove self");
+        }
         return this.tenantService.removeMember(tenantId, user);
     }
 
     @Put('/:tenantId/member/:email/scope')
-    @UseGuards(JwtAuthGuard, ScopeGuard)
-    @Scopes(ScopeEnum.TENANT_ADMIN)
+    @UseGuards(JwtAuthGuard)
     async updateScope(
         @Request() request,
         @Param('tenantId') tenantId: string,
@@ -95,7 +94,7 @@ export class MemberController {
     ): Promise<Scope[]> {
         const user = await this.usersService.findByEmail(email);
         let tenant = await this.tenantService.findById(tenantId);
-        await this.securityService.currentUserShouldBeTenantAdmin(request, tenant.domain)
+        this.securityService.check(request, Action.Update, tenant);
         return this.tenantService.updateScopeOfMember(body.scopes, tenantId, user);
     }
